lib: move WebAssembly Web API into separate file
Refs: https://github.com/nodejs/node/pull/42960#discussion_r864871357
