+++ lib/internal/bootstrap/pre_execution.js
@@ -5,7 +5,6 @@ const {
   ObjectDefineProperties,
   ObjectDefineProperty,
   ObjectGetOwnPropertyDescriptor,
-  PromiseResolve,
   SafeMap,
   SafeWeakMap,
   StringPrototypeStartsWith,
@@ -26,9 +25,7 @@ const {
 
 const { Buffer } = require('buffer');
 const {
-  ERR_INVALID_ARG_TYPE,
   ERR_MANIFEST_ASSERT_INTEGRITY,
-  ERR_WEBASSEMBLY_RESPONSE,
 } = require('internal/errors').codes;
 const assert = require('internal/assert');
 
@@ -222,49 +219,8 @@ function setupFetch() {
   });
 
   // The WebAssembly Web API: https://webassembly.github.io/spec/web-api
-  internalBinding('wasm_web_api').setImplementation((streamState, source) => {
-    (async () => {
-      const response = await PromiseResolve(source);
-      if (!(response instanceof lazyUndici().Response)) {
-        throw new ERR_INVALID_ARG_TYPE(
-          'source', ['Response', 'Promise resolving to Response'], response);
-      }
-
-      const contentType = response.headers.get('Content-Type');
-      if (contentType !== 'application/wasm') {
-        throw new ERR_WEBASSEMBLY_RESPONSE(
-          `has unsupported MIME type '${contentType}'`);
-      }
-
-      if (!response.ok) {
-        throw new ERR_WEBASSEMBLY_RESPONSE(
-          `has status code ${response.status}`);
-      }
-
-      if (response.bodyUsed !== false) {
-        throw new ERR_WEBASSEMBLY_RESPONSE('body has already been used');
-      }
-
-      if (response.url) {
-        streamState.setURL(response.url);
-      }
-
-      // Pass all data from the response body to the WebAssembly compiler.
-      const { body } = response;
-      if (body != null) {
-        for await (const chunk of body) {
-          streamState.push(chunk);
-        }
-      }
-    })().then(() => {
-      // No error occurred. Tell the implementation that the stream has ended.
-      streamState.finish();
-    }, (err) => {
-      // An error occurred, either because the given object was not a valid
-      // and usable Response or because a network error occurred.
-      streamState.abort(err);
-    });
-  });
+  const { wasmStreamingCallback } = require('internal/wasm_web_api');
+  internalBinding('wasm_web_api').setImplementation(wasmStreamingCallback);
 }
 
 // TODO(aduh95): move this to internal/bootstrap/browser when the CLI flag is
